python/scripts/webp_to_png_converter/requirements.txt
Pillow>=9.0.0 # Para manejo de imágenes
